backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
boto3==1.40.17
botocore==1.40.17
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.6
distro==1.9.0
Django==5.2.5
django-cors-headers==4.2.0
django-storages==1.14.6
djangorestframework==3.16.1
dnspython==2.7.0
edge-tts==7.2.1
eventlet==0.40.2
frozenlist==1.7.0
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jmespath==1.0.1
josepy==2.1.0
kombu==5.5.4
lxml==6.0.1
mozilla-django-oidc==4.0.1
multidict==6.6.4
openai==1.102.0
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
propcache==0.3.2
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-pptx==1.0.2
redis==6.4.0
requests==2.32.5
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tabulate==0.9.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
xlsxwriter==3.2.5
yarl==1.20.1